Added a fixme and a temporary fix for deadlock (see this commit's extended message for more info).
This deadlock happens in a scenario where two consecutive NET messages (NET(t1) and NET(t2)) are sent by a federate f without receiving a TAG or PTAG in between. t1 can be smaller than t2, and NET(t2) will overwrite t1 as the recorded next event for f in the RTI. If f is granted a PTAG for t1 and a TAG for t2, it can block on t1 forever because it has received a PTAG(t1), without receiving a TAG(t1) after.