[scheduler] More relatively minor tweaks.
I run the risk of overfitting by doing this. However, it makes sense not to do a state change before collecting any information about the initial state -- the initial state was chosen for a reason. The decision to make left/right transition probabilities more even was not really based on empirical results as much as a vague sort of common sense.