[benchmarks] Set timeout.
This timeout is _much_ lower than what we previously had to use.
This is because of the size=fast parameter which cuts down the
problem size for benchmarks that take unreasonably long to execute
(e.g., SortedLinkedList). It is also an order of magnitude greater
than the max execution time using the unthreaded C runtime, so any
implementation that times out is unacceptably slow.