Fix flaky timedwait on Windows
The Windows implementation of `lf_timed_wait` calculates between the timepoint
to wait until and the current physical time. In some seldom cases, the current physical
time could already be larger then the specified timepoint. This will lead to a
negative delay to `SleepConditionVariableCS`. While the documentation does not
specify this precisely, experimentation showed that negative delays are
interpreted as infinite delays and hence the program would block forever.

This fix makes to changes. First, it checks for negative delays and returns
without waiting if the delay is less or equal to zero. Second, the number of
milliseconds tow wait for is rounded up in order to prevent unnecessary busy
waiting when the call to `SleepConditionVariableCS` returns early.