Remove _instant_t.
There might have been a reason why this was created, but I consider it
to be excess generality. I do not see us moving toward representing time
with less than 64 bits in the near future unless I missed something in
our conversations on the topic, and I question whether emulation of 64
bit integers is as much slower than the alternatives as it is imagined
to be. It has also been commented multiple times that defining
_instant_t in one place and then always aliasing it as instant_t smells
like overengineering.