Do not decrement the tag barrier if it is zero.
This fix seems dangerous and looks like it can cause deadlock if
increments and decrements are ordered incorrectly. However, I do not
think that it will cause deadlock because it seems like the federate/RTI
communication protocol is supposed to ensure that the increment happens
first, if it happens at all. Furthermore, I do not see a way around a
fix like this because the logic for deciding whether or not to increment
the tag barrier seems nontrivial.