Fix another deadlock.
After this change I was able to run Python's StopAtShutdown 95 times
without observing any deadlock, whereas before I could consistently
reproduce the deadlock (when using the macOS runner from CI) by running
it several times.

It is not entirely clear what this deadlock has to do with the changes
in this branch. Although I have not reproduced it in master (I have
tried, and after running it 95 times I did not see deadlock), it does
seem like a programming mistake to allow the stop tag to increase.
Besides, the error was very flaky, non-reproducible on Ubuntu, and quite
possibly present even if it is not observed.