Remove the "waiting for tag" field.
When a federate sends a NET for a certain tag, and it has upstream
federates, we incorrectly assumed that the federate needs to receive a
TAG or PTAG in order to allow it to proceed to that NET.

This assumption is incorrect because after sending the NET, the federate
might receive a message from an upstream federate; that message might
lead to a newly lowered NET.

The bug that resulted from the incorrect assumption is that the federate
would not proceed to execute the event enabled by the message it
received after sending the NET. This is because the federate does not
realize that it has anything to do; it still thinks that it cannot do
anything until it receives a TAG or PTAG that enables it to process the
NET. Therefore, the federation deadlocks.

The fix is to compute a new next event tag whenever the event queue
changes.