Partially revert some implementation changes.
I think that the new implementation incorrectly handled the case for
delay_strict when the interval is zero and a single-microstep delay is
introduced. This highlights the amount of subtlety in the meaning of the
function's behavior and shows that it is not as simple as returning the
latest tag that is less than the sum. Instead, it has to return the
latest tag in the current interval at which nothing can happen provided
that nothing happens at the current tag, like I explained originally.