Account for federate starting near RTI start time.
This is arguably overfitting to the behavior of the build script. We can
revisit it if an when it comes time to merge. The issue is that if the
federate starts at about the same time as the RTI, it is racing: it
might try the port before the RTI has set up its server. We want a short
initial retry interval so that tests can run within a few milliseconds.