Move API calls around.
Apparently if the tokio is accessed after the exit sequence has already
begun, it crashes...? Because somehow the destruction of the whole
runtime starts _before_ the termination sequence that is configured
using `atexit`, even if `atexit` was called after the tokio runtime was
started up, and even, I think, after the .so was loaded (I am less sure
about this)? I do not know what is going on under the hood here.