Update RTI dockerfile to match file layout change
This could harm reproducibility and could increase the size of the RTI
image. However, reproducibility is actually not much worse than the
status quo which was still copying whole directories. Also, the amount
of extra stuff copied is small. In the long run this will keep us from
running into further problems as we continue to refactor this repo
without testing the dockerized RTI build in CI (I am assuming that it
is not worthwhile to add a test for it in CI)